Sprint-2-Financial-Advisor/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chromadb==1.0.8
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.5
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.7
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
durationpy==0.9
emoji==2.14.1
executing==2.0.1
fastapi==0.115.9
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.55.8
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2023.10.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
google-pasta==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
grandalf==0.8
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inquirerpy==0.3.4
ipykernel==6.28.0
ipython==8.19.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.6.1
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.2
keras==3.8.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.1.147
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
lightning==2.5.0.post0
lightning-utilities==0.12.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.4.1
mmh3==5.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
multitasking==0.0.11
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.38.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.21.1
openai==1.78.1
opencv-python==4.11.0.86
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
patsy==1.0.1
peewee==3.17.9
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
posthog==4.0.1
preshed==3.0.9
prometheus_client==0.21.1
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==5.29.3
psutil==5.9.7
pure-eval==0.2.2
pyarrow==19.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.8.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.17.2
PyMuPDF==1.26.0
pyparsing==3.2.1
pypdf==5.6.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-json-logger==3.2.1
python-magic==0.4.27
python-oxmsg==0.0.2
python-slugify==8.0.4
pytorch-lightning==2.5.0.post0
pytz==2025.1
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==25.1.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.45.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.5.10
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.6.1
torchvision==0.21.0+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.1
transformers==4.48.2
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
unstructured==0.17.2
unstructured-client==0.36.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.12
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.54
zipp==3.21.0
zstandard==0.23.0
